moved TAC code
